server: Port to Windows using mingw-w64.
This allows nbdkit to be cross-compiled for Windows (from Linux).  On
Fedora[1] you would use the commands below:

  sudo dnf install mingw64-gcc mingw64-dlfcn \
                   mingw64-gnutls mingw64-xz mingw64-zlib
  mingw64-configure --disable-plugins
  make

This results in a Windows binary which can be run on Windows or on
Linux under Wine.

It also probably allows nbdkit to be compiled on Windows natively
using the mingw-w64 port of GCC, but I did not test that.

There are several limitations compared to running nbdkit on a real
server, see ‘TODO’ for the complete list.

[1] Note that mingw64-configure is a Fedora-specific script which runs
./configure with extra parameters to make cross-compilation for
Windows work.  You can add other configure parameters to the command
or if you're not using Fedora then use ‘./configure --host=...’
instead.