Add linuxdisk plugin.
This plugin allows you to create a complete ext2, ext3 or ext4
filesystem in a GPT partitioned disk image.  This can be attached as a
disk to a Linux virtual machine.  It is implemented using e2fsprogs
mke2fs ‘-d’ option thus allowing the implementation to be very small
and simple, with all the hard work done by mke2fs.

Although there is some overlap with nbdkit-iso-plugin and
nbdkit-floppy-plugin, the implementations and use cases of all three
plugins are sufficiently different that it seems to make sense to add
another plugin rather than attempting to extend one of the existing
plugins.

Largely to avoid user error this plugin is read-only.  This is a major
difference from the floppy plugin: that plugin allows files to be
modified (but not resized or created) and writes those changes through
to the backing filesystem.  While this plugin could easily be made
writable, this would cause almost certain disk corruption when someone
connected two clients at the same time.  In any case it doesn't make
much sense for it to be writable by default since the expectation that
writes would somehow modify the original directory on the host
filesystem cannot be satisfied by this or any reasonable
implementation.  Users can add the cow filter on top if they really
want writes and know what they are doing: instructions plus disclaimer
about this are included in the man page.

Eventually we could replace the supermin libext2fs code with this
plugin, but there are some missing features that would need to be
implemented first.