server: Add libnbdkit.so.
This essentially turns the whole of nbdkit into a library.  There is a
rump nbdkit program left which simply contains a main() function that
forwards to the private _nbdkit_main() function in the library.

The reason for this is to allow nbdkit to be compiled on Windows,
because Windows does not allow shared libraries to contain any
undefined symbols.  nbdkit previously relied on undefined symbols in
all plugins and filters (eg. nbdkit_parse_int) which get resolved at
run time by the nbdkit program.  In order to make this work on Windows
we need to have a new library (libnbdkit.so on Linux, but probably
called something like LIBNBDKIT.DLL on Windows) which will contain
these symbols, and plugins can then be compiled with -lnbdkit so they
will link with this library, hence no undefined symbols.

Note this change is backwards compatible.  Because the rump nbdkit
binary links with -lnbdkit it still contains and reexports all the
symbols, so plugins and filters which don't explicitly link with
-lnbdkit will still be loadable (on existing platforms).